Use EnumIndexedVector in J2KEncoder.

+++ src/lib/j2k_encoder.cc
@@ -229,7 +229,7 @@ J2KEncoder::encode (shared_ptr<PlayerVideo> pv, DCPTime time)
 		/* This frame already has J2K data, so just write it */
 		_writer->write (pv->j2k(), position, pv->eyes ());
 		frame_done ();
-	} else if (_last_player_video[static_cast<int>(pv->eyes())] && _writer->can_repeat(position) && pv->same (_last_player_video[static_cast<int>(pv->eyes())])) {
+	} else if (_last_player_video[pv->eyes()] && _writer->can_repeat(position) && pv->same(_last_player_video[pv->eyes()])) {
 		LOG_DEBUG_ENCODE("Frame @ %1 REPEAT", to_string(time));
 		_writer->repeat (position, pv->eyes ());
 	} else {
@@ -250,7 +250,7 @@ J2KEncoder::encode (shared_ptr<PlayerVideo> pv, DCPTime time)
 		_empty_condition.notify_all ();
 	}
 
-	_last_player_video[static_cast<int>(pv->eyes())] = pv;
+	_last_player_video[pv->eyes()] = pv;
 	_last_player_video_time = time;
 }
